refactor(snak): snack extracted to his own container
