+++ src/discrete-layer/views/discrete-layer-view.tsx
@@ -2,88 +2,39 @@ import React, { useState, useEffect } from 'react';
 import {
   Box,
   CesiumWMTSLayer,
-  RCesiumWMTSLayerOptions,
-  RCesiumWMSLayerOptions,
   CesiumWMSLayer,
-  RCesiumXYZLayerOptions,
   CesiumXYZLayer,
   CesiumOSMLayer,
-  RCesiumOSMLayerOptions,
 } from '@map-colonies/react-components';
 import { observer } from 'mobx-react-lite';
-import { Button, Drawer, DrawerContent, DrawerHeader, DrawerSubtitle, DrawerTitle, Snackbar, SnackbarAction } from '@map-colonies/react-core';
+import { Button, Drawer, DrawerContent, DrawerHeader, DrawerSubtitle, DrawerTitle } from '@map-colonies/react-core';
 import { DateTimeRangePickerFormControl, SupportedLocales } from '@map-colonies/react-components';
 import { FormattedMessage, useIntl } from 'react-intl';
-import { ResponseState } from '../../common/models/response-state.enum';
 import CONFIG from '../../common/config';
+import { MOCK_DATA_IMAGERY_LAYERS_ISRAEL } from '../../__mocks-data__/search-results.mock';
+import { osmOptions, wmsOptions, wmtsOptions, xyzOptions } from '../../common/helpers/layer-options';
 import { useStore } from '../models/rootStore';
 import { MapContainer } from '../components/map-container';
 import { DrawerOpener } from '../components/drawer-opener/drawer-opener';
 import { LayersResultsComponent } from '../components/layers-results/layers-results';
 import './discrete-layer-view.css';
-import { MOCK_DATA_IMAGERY_LAYERS_ISRAEL } from '../../__mocks-data__/search-results.mock';
-
-type ServerType = 'geoserver' | 'carmentaserver' | 'mapserver' | 'qgis';
-
-/* eslint-disable */
-const wmtsOptions: RCesiumWMTSLayerOptions = {
-  url: CONFIG.WMTS_LAYER.URL,
-  layer: CONFIG.WMTS_LAYER.LAYER,
-  style: CONFIG.WMTS_LAYER.STYLE,
-  format: CONFIG.WMTS_LAYER.FORMAT,
-  tileMatrixSetID: CONFIG.WMTS_LAYER.TILE_MATRIX_SET_ID,
-  maximumLevel: CONFIG.WMTS_LAYER.MAXIMUM_LEVEL,
-};
-
-const wmsOptions: RCesiumWMSLayerOptions = {
-  url: CONFIG.WMS_LAYER.URL,
-  layers: CONFIG.WMS_LAYER.PARAMS.LAYERS,
-};
-
-const xyzOptions: RCesiumXYZLayerOptions = {
-  url: CONFIG.XYZ_LAYER.URL,
-};
-
-const osmOptions: RCesiumOSMLayerOptions = {
-  url: CONFIG.OSM_LAYER.URL,
-}
-/* eslint-enable */
 
 const tileOtions = { opacity: 0.5 };
 
 const mapActionsWidth = '400px';
 
-interface SnackDetails {
-  message: string;
-}
-
 const DiscreteLayerView: React.FC = observer(() => {
   const { discreteLayersStore } = useStore();
-  const [snackOpen, setSnackOpen] = useState(false);
   const [resultsOpen, setResultsOpen] = useState(false);
   const [filtersOpen, setFiltersOpen] = useState(false);
-  const [snackDetails, setSnackDetails] = useState<SnackDetails>({
-    message: '',
-  });
   const intl = useIntl();
-  useEffect(() => {
-    switch (discreteLayersStore.state) {
-      case ResponseState.ERROR:
-        setSnackOpen(true);
-        setSnackDetails({
-          message: 'snack.message.failed',
-        });
-        break;
-      case ResponseState.DONE:
-        setSnackOpen(true);
-        setSnackDetails({
-          message: 'snack.message.success',
-        });
-        break;
-      default:
-        break;
-    }
-  }, [discreteLayersStore.state]);
+
+  // TODO REMOVE: EXAMLPE HOW TO TRIGGER SNACK  
+  useEffect(()=>{
+    setTimeout(()=>{
+      discreteLayersStore.getLayersImages();
+    }, 7000); 
+  },[]);
 
   return (
     <MapContainer
@@ -155,20 +106,6 @@ const DiscreteLayerView: React.FC = observer(() => {
             onClick={setResultsOpen}
           />
 
-          {!!snackDetails.message && (
-            <Snackbar
-              open={snackOpen}
-              onClose={(evt): void => setSnackOpen(false)}
-              message={intl.formatMessage({ id: snackDetails.message })}
-              dismissesOnAction
-              action={
-                <SnackbarAction
-                  label={intl.formatMessage({ id: 'snack.dismiss-btn.text' })}
-                  onClick={(): void => console.log('dismiss clicked')}
-                />
-              }
-            />
-          )}
       </>,
       ]}
       mapContent={
